fix(kernel): revert ROL64 to rotate(long,long) for GCN; make gcn_s4_mode optional
- cosmic_harmony_deeksha.cl: ROL64 changed back from rotate(ulong,ulong)
  (introduced in 0ccd21d7) to rotate(long,long) which was the original
  AMD Vega/GCN compiler-bug workaround.
- gpu_backend.rs: gcn_s4_mode is no longer mandatory.  The miner now tries
  the full GPU pipeline first on GCN/RDNA devices.  s4-only mode is only
  activated when ZION_GCN_S4_MODE=1 is set explicitly.
- This lets us test whether the rotate fix resolves the s4_memhard
  mismatch on gfx900 without forcing the low-hashrate workaround.

+++ V3/L1/cosmic-harmony/src/gpu/kernels/cosmic_harmony_deeksha.cl
@@ -39,10 +39,9 @@
 #define WGS 64
 #endif
 
-/* AMD compiler bug workaround: 64-bit rotate by 8 is miscompiled
- * for some GCN/RDNA driver combos. Use OpenCL built-in rotate
- * with ulong (unsigned) which is handled correctly on all targets. */
-#define ROL64(x, n) rotate((ulong)(x), (ulong)(n))
+/* AMD Vega / GCN compiler bug: 64-bit rotate by 8 is miscompiled.
+ * Use OpenCL built-in rotate(long,long) which is handled correctly. */
+#define ROL64(x, n) rotate((long)((ulong)(x)), (long)((ulong)(n)))
 #define XTIME(a) ((uchar)(((a) << 1) ^ ((((a) >> 7) & 1) * 0x1b)))
 
 /* Chi macro: one 5-element row, no temp array (from optimized v3 kernel) */
